Directional Shooting
Vehicle fires in all 8 directions. Need to add a rotation and spawn
points for diag and right facing positions to switch between as the
weapon rotates. Currently the bullets fire in the correct directions
however the bullets themselves are continuously facing up and spawn away
from weapon tip when turned.