Controller Support
Modified the vehicle script to allow a controller to be used in the
game. Primarily the DPAD. Up, Down, Left, Right, UpRight, UpLeft
DownRight, and DownLeft are all supported. The object is turned a
specific degree to emulate the look of older nes games like Jackal. Now
that that is supported I need to add a dial-like control pattern to turn
the car appropriately. If the car is facing up and the play presses
down, the car must rotate on its own the new desided heading. Also other
buttons will be added. I wanted this up in case my PC exploded. It was
difficult to get the way I wanted but I am happy with the results. Some
times you just need to sleep on a problem ; )